requirements.txt
Flask
Flask-Bcrypt
mysql-connector-python
requests